backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "pydantic >1.4",
    "betterlogging >=1.0.0,<2.0.0",
    "pydantic-settings >=2.10.1,<3.0.0",
    "aiohttp-socks >=0.10.1,<1.0.0",
    "aiogram >=3.21.0,<4.0.0",
    "aiogram-dialog >=2.4.0,<3.0.0",
    "ruff >=0.12.4,<1.0.0",
]
